// src/common/constants/enums.ts
export enum NftSourceEnum {
  UNIVERSE = 'universe',
  SCRAPER = 'scraper',
}
export enum NftCollectionSourceEnum {
  UNIVERSE = 'universe',
  SCRAPER = 'scraper',
}

export enum NftStatusEnum {
  SAVED = 'saved',
  MINTING = 'minting',
  MINTED = 'minted',
}
export enum NftCollectionStatusEnum {
  SAVED = 'saved',
  DEPLOYING = 'deploying',
  DEPLOYED = 'deployed',
}

export enum MetadataStorageEnum {
  ONCHAIN = 'onchain',
  OFFCHAIN = 'offchain',
}

export enum NftTransactionStatusEnum {
    PENDING = 'pending',
    FAILED = 'failed',
    SUCCESS = 'success',
}